refactor: Move `SMJ` tests into own file
