+++ datafusion/physical-plan/src/joins/sort_merge_join/mod.rs
@@ -22,3 +22,6 @@ pub use exec::SortMergeJoinExec;
 mod exec;
 mod metrics;
 mod stream;
+
+#[cfg(test)]
+mod tests;
